# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-commerce API",
    "version": "1.0.0",
    "description": "API documentation for the AI Shopping Agent e-commerce platform"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["name", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Successful response with JWT token" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login to get a JWT token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Successful response with JWT token" }
        }
      }
    },
    "/products/search": {
      "get": {
        "summary": "Search products",
        "parameters": [
          { "name": "q", "in": "query", "schema": { "type": "string" } },
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "min_price", "in": "query", "schema": { "type": "number" } },
          { "name": "max_price", "in": "query", "schema": { "type": "number" } },
          { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["price_asc", "price_desc", "rating"] } },
          { "name": "in_stock", "in": "query", "schema": { "type": "boolean" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": { "description": "Successful response" }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "summary": "Get product by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Successful response" }
        }
      }
    },
    "/products/{id}/reviews": {
      "get": {
        "summary": "Get product reviews",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Successful response" }
        }
      }
    },
    "/categories": {
      "get": {
        "summary": "Get all categories",
        "responses": {
          "200": { "description": "Successful response" }
        }
      }
    },
    "/cart": {
      "get": {
        "summary": "Get current user's cart",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Successful response" }
        }
      }
    },
    "/cart/add": {
      "post": {
        "summary": "Add item to cart",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "product_id": { "type": "string" },
                  "quantity": { "type": "integer", "minimum": 1 }
                },
                "required": ["product_id", "quantity"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Successful response" }
        }
      }
    },
    "/cart/{item_id}": {
      "put": {
        "summary": "Update cart item quantity",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "item_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quantity": { "type": "integer", "minimum": 1 }
                },
                "required": ["quantity"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Successful response" }
        }
      },
      "delete": {
        "summary": "Remove item from cart",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "item_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Successful response" }
        }
      }
    },
    "/checkout": {
      "post": {
        "summary": "Checkout current cart",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "address": { "type": "string" },
                  "payment_method": { "type": "string", "default": "Card" }
                },
                "required": ["address"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Successful response" }
        }
      }
    },
    "/orders": {
      "get": {
        "summary": "Get current user's orders",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Successful response" }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "summary": "Get order by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Successful response" }
        }
      }
    }
  }
}